+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## v0.0.30 "Selected Recordings :rarum IX"
+Refactored "Integrations" into separate Import and Export code and configs
+Changed -i parameter to -f to repurpose -i for something else in the next version
+
+
 ## v0.0.29 "One Quiet Night"
 FluentD integration